Add UpdateFollowUpSequenceService

+++ src/modules/followUp/services/DeleteFollowUpSequenceService.ts
@@ -1,4 +1,5 @@
 import { injectable, inject } from 'tsyringe';
+import AppError from '@shared/errors/AppError';
 
 import IFollowUpSequenceRepository from '@modules/followUp/repositories/IFollowUpSequenceRepository';
 
@@ -9,11 +10,18 @@ interface IRequest {
 @injectable()
 class DeleteFollowUpSequenceService {
   constructor(
-    @inject('FollowUpSequencesRepository')
+    @inject('FollowUpSequenceRepository')
     private emailModelRepository: IFollowUpSequenceRepository,
   ) {}
 
   public async execute({ id }: IRequest): Promise<void> {
+    const followUpSequence = await this.emailModelRepository.findById(id);
+
+    if (!followUpSequence) throw new AppError('Follow up sequence not found.');
+
+    if (followUpSequence?.title === 'Padrão')
+      throw new AppError('Cannot delete default follow up sequence.');
+
     await this.emailModelRepository.delete(id);
   }
 }
